add slim variant for cista

+++ definitions/graph.capnp
@@ -2,13 +2,13 @@
 
 struct Graph {
   struct Edge {
-    from @0 :UInt32;
-    to @1 :UInt32;
-    weight @2 :UInt32;
+    from @0 :UInt16;
+    to @1 :UInt16;
+    weight @2 :UInt16;
   }
 
   struct Node {
-    id @0 :UInt32;
+    id @0 :UInt16;
     name @1 :Text;
     outEdges @2 :List(Edge);
     inEdges @3 :List(Edge);
